weight

+++ app/src/main/java/com/tanya/health_care/ChangeMyFoodFragment.java
@@ -111,7 +111,7 @@ public class ChangeMyFoodFragment extends Fragment {
                     }
 
                     try {
-                        float weightValue = Float.parseFloat(weightText);
+                        int weightValue = Integer.parseInt(weightText);
                         int caloriesValue = Integer.parseInt(caloriesText);
                         int proteinValue = Integer.parseInt(proteinText);
                         int fatValue = Integer.parseInt(fatText);
